backend/requirements.txt
Flask==3.1.3
python-dotenv==1.2.2
pdfplumber==0.11.9
google-generativeai==0.8.6
gunicorn==23.0.0